send initial partial mesages
