+++ server/constants.js
@@ -24,7 +24,7 @@ exports.WEBSOCKET_CHANNEL = (topic, symbolOrUserId) => {
 		case 'deposit':
 			return `deposit:${symbolOrUserId}`;
 		case 'chat':
-			return `chat:${symbolOrUserId}`;
+			return 'chat';
 		default:
 			return;
 	}
